added notifications

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -47,11 +47,32 @@ export async function POST(
     if (!loggedInUser)
       return Response.json({ error: "Unauthorized" }, { status: 401 });
 
-    await prisma.like.upsert({
-      where: { userId_postId: { userId: loggedInUser.id, postId } },
-      create: { userId: loggedInUser.id, postId },
-      update: {},
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: { userId: true },
     });
+    if (!post)
+      return Response.json({ error: "Post not found" }, { status: 404 });
+
+    await prisma.$transaction([
+      prisma.like.upsert({
+        where: { userId_postId: { userId: loggedInUser.id, postId } },
+        create: { userId: loggedInUser.id, postId },
+        update: {},
+      }),
+      ...(loggedInUser.id !== post.userId
+        ? [
+            prisma.notification.create({
+              data: {
+                issuerId: loggedInUser.id,
+                receipientId: post.userId,
+                postId,
+                type: "LIKE",
+              },
+            }),
+          ]
+        : []),
+    ]);
 
     return new Response();
   } catch (error) {
@@ -69,9 +90,26 @@ export async function DELETE(
     if (!loggedInUser)
       return Response.json({ error: "Unauthorized" }, { status: 401 });
 
-    await prisma.like.deleteMany({
-      where: { userId: loggedInUser.id, postId },
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: { userId: true },
     });
+    if (!post)
+      return Response.json({ error: "Post not found" }, { status: 404 });
+
+    await prisma.$transaction([
+      prisma.like.deleteMany({
+        where: { userId: loggedInUser.id, postId },
+      }),
+      prisma.notification.deleteMany({
+        where: {
+          issuerId: loggedInUser.id,
+          receipientId: post.userId,
+          postId,
+          type: "LIKE",
+        },
+      }),
+    ]);
 
     return new Response();
   } catch (error) {
